minor changes (conversions)

+++ lab2/utils.c
@@ -4,13 +4,13 @@
 
 int(util_get_LSB)(uint16_t val, uint8_t *lsb) {
     if (lsb == NULL) return 1;
-    *lsb = (uint8_t)val;
+    *lsb = val;
     return 0;
 }
 
 int(util_get_MSB)(uint16_t val, uint8_t *msb) {
     if (msb == NULL) return 1;
-    *msb = (uint8_t)(val >> 8);
+    *msb = (val >> 8);
     return 0;
 }
 
@@ -18,6 +18,6 @@ int (util_sys_inb)(int port, uint8_t *value) {
     if(value == NULL) return 1;
     uint32_t n = 0;
     if(sys_inb(port, &n)) return 1;
-    *value = (uint8_t)n;
+    *value = n;
     return 0;
 }
